fix: background debounced
